+++ lib/app/services/converters/address_model_parsing.dart
@@ -1,0 +1,13 @@
+part of '../converters.dart';
+
+extension AddressModelParsing on AddressModel {
+  PoiAddress transformToPoiAddress() {
+    final address = this;
+    return PoiAddress(
+      city: address.city.isEmpty ? null : address.city,
+      zip: address.zipCode.isEmpty ? null : address.zipCode,
+      street: address.street.isEmpty ? null : address.street,
+      houseNumber: address.houseNumber.isEmpty ? null : address.houseNumber,
+    );
+  }
+}
