+++ lib/app/services/converters/lat_lng_parsing.dart
@@ -1,0 +1,22 @@
+part of '../converters.dart';
+
+extension LatLngParsing on LatLng {
+  String toLngLatString() {
+    return transformToGeoJsonCoords().join(',');
+  }
+
+  List<double> transformToGeoJsonCoords() {
+    return [longitude, latitude];
+  }
+
+  List<double> transformToGeoJsonBBox(LatLng northEast) {
+    final southWest = this;
+    final coords = southWest.transformToGeoJsonCoords();
+    coords.addAll(northEast.transformToGeoJsonCoords());
+    return coords;
+  }
+
+  String transformToGeoJsonBBoxString(LatLng northEast) {
+    return transformToGeoJsonBBox(northEast).join(',');
+  }
+}
