+++ lib/app/services/converters/lat_lng_parsing_extended.dart
@@ -1,0 +1,8 @@
+part of '../converters.dart';
+
+extension LatLngParsingExtended on List<double> {
+  LatLng transformToLatLng() {
+    if (length != 2) throw ArgumentError('coordinates should contain 2 items');
+    return LatLng(this[1], this[0]);
+  }
+}
