+++ lib/app/services/converters/poi_type_parsing.dart
@@ -1,0 +1,16 @@
+part of '../converters.dart';
+
+extension PoiTypeParsing on PoiType {
+  PoiServiceType transformToPoiServiceType() {
+    switch (this) {
+      case PoiType.flyerSpot:
+        return PoiServiceType.flyer;
+      case PoiType.poster:
+        return PoiServiceType.poster;
+      case PoiType.house:
+        return PoiServiceType.door;
+      case PoiType.swaggerGeneratedUnknown:
+        throw UnimplementedError();
+    }
+  }
+}
