+++ lib/app/services/converters/poster_status_parsing.dart
@@ -1,0 +1,12 @@
+part of '../converters.dart';
+
+extension PosterStatusParsing on PosterStatus {
+  PoiPosterStatus transformToPoiPosterStatus() {
+    return switch (this) {
+      PosterStatus.ok => PoiPosterStatus.ok,
+      PosterStatus.damaged => PoiPosterStatus.damaged,
+      PosterStatus.missing => PoiPosterStatus.missing,
+      PosterStatus.removed => PoiPosterStatus.removed,
+    };
+  }
+}
