+++ lib/features/campaigns/screens/mixins.dart
@@ -1,0 +1,14 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:gruene_app/app/services/nominatim_service.dart';
+import 'package:gruene_app/app/theme/theme.dart';
+import 'package:gruene_app/i18n/translations.g.dart';
+import 'package:motion_toast/motion_toast.dart';
+
+part 'mixins/address_extension.dart';
+part 'mixins/confirm_delete.dart';
+part 'mixins/door_validator.dart';
+part 'mixins/flyer_validator.dart';
+part 'mixins/focus_area_info.dart';
+part 'mixins/poster_validator.dart';
